Make correction in "Sigils"
The macro `sigil_r` was incorrectly identified as a function. One
could have merely changed the word 'function' to 'macro'. However,
the revelation that it is a macro and not a function could pose a
distraction and lead to confusion ahead of the final paragraph of
the page. Consequently, I chose to remove the word 'function'.