Indent with 4 spaces instead of using ticks
There are places in documentation where ticks apply (such as code that
has to be highlighted), and there are places that don't have to be
highlighted.

There's no need to keep everything in sync and wrap both cases with
ticks.

With this commit blocks that don't need any code highlighting are
indented with 4 spaces, as they initially were.