Update description for adding run method
IMO current version of the tutorial does not make clear a new method must be added. I stumbled upon this and replaced 
```
  def run({:create, bucket}) do
    KV.Registry.create(KV.Registry, bucket)
    {:ok, "OK\r\n"}
  end
```

with

```
  def run({:create, bucket}, pid) do
    KV.Registry.create(pid, bucket)
    {:ok, "OK\r\n"}
  end
```
instead adding another "version" of the method. This confusion could come with OOP way of thinking. I am not a native speaker, so the wording could be changed.