Two small edits
An attempt to simplify and add clarity to the text. (I'm an Elixir newbie, but I took a gamble on changing "the sides can't match" to "the sides can't be matched". Based on what I've understood from Dave Thomas's book (the first few pages), I think my version is clearer. Of course I may be wrong!)