Adding title attribute to embedded YouTube video
Using the title of the video itself. Rocket Validator also flags the
`frameborder` attribute as a problem *but* it is still present in the
embed code straight from YouTube. I’m going to assume the YouTube team
continues to include that attribute in 2017 for well considered reasons
and leave it as is for now.