Rocket validator (#962)
* Image element tweaks

Adding alt attribute, removing px unit from width attribute.

* Removing chrome=1 from X-UA-Compatible meta

Relevant discussion behind the validator error seems to be here:
http://lists.w3.org/Archives/Public/www-validator/2011Mar/0053.html

* Adding lang attribute to html element

Per Rocket Validator flagging the lack of a lang element as a serious
accessibility issue. *But* this could pose problems for any
internationalization strategy that would involve publishing multiple
translations under a single top level include.

* Adding aria-label attributes for site search widget

See
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Tec
hniques/Using_the_aria-label_attribute

* Adding title attribute to embedded YouTube video

Using the title of the video itself. Rocket Validator also flags the
`frameborder` attribute as a problem *but* it is still present in the
embed code straight from YouTube. I’m going to assume the YouTube team
continues to include that attribute in 2017 for well considered reasons
and leave it as is for now.