Clear ambiguity in race conditions
Former text could give misleading impression on how to resolve interprocess communication problems.

I am also not sure whether the test "removes bucket on crash" behaves deterministically.
Agent.stop/2 is a sync method, which means that by the time the shell receives control the bucket is down; but can we be sure that the registry receives the message about bucket being `:DOWN` before it receives the message `:create`?  If it was deterministic messages should be sent in the following order:
1. Shell sends `:shutdown` to the bucket;
2. VM sends `:DOWN` message to the registry;
3. Shell sends `:create` message to registry.
Above order ensures that the test is deterministic.

I have not found the proper documentation to support my claim, but can't the order be as following?
1. Shell sends `:shutdown` to the bucket;
2/3. VM sends `:DOWN` message to the registry;
2/3. Shell sends `:create` message to registry.
Above order means that we are not sure which message reaches registry's inbox first - `:create:` or `:DOWN`.

I think it requires further investigation or a clear answer.