comparing named/anonymous functions
I'm not so sure of this edit, but it does make more sense to me this way.  I might be misunderstanding radically though.  

I've added a back-link to the place where anonymous functions had been introduced, and named functions mentioned.

I state (please correct if wrong) that there is no difference in the potential of anonymous vs. named functions.

I explain the capture operator `&` as a sort of function anonymiser.  too many words?  too few?  it does help my understanding, this way.