please review
I did not yet process the feedback regarding the `reason`.

added license and copyright statements, they are more explicit
in the source file than in the pdf.

btw: using the global header, because they are page-wide and
occupy less space this way.

added an 'applies globally' title to the `reason` and
`then_what` explanations.  also changed the background colour.

are there extra patterns against which to match the `then_what`
value? as it stands now it looks a bit ugly, or we have to
choose a shorter name.

I didn't look into making the itemized list more compact.