Add upcoming conferences section and add CodeBeam SF & ElixirConf EU
Added in important links; over the community section.

I have made it such that it is easy to add and remove conferences from
a list; I thought it would be best to order the conferences such that
the most recent is at the top, and the template include room for a
logo, the location, and a little space to write some information about
dates.