Clarify how agent passes state to anonymous func (#1380)
When encountering Agents for the first time, it is not clear that they pass their state to anonymous functions in this way. This problem stumped me for a whole day yesterday, when I couldn't figure out what was going on with that `&1`. 

It finally made sense when someone on the ElixirForum gave me this crucial piece of information, that the Agent calls the anonymous function and passes its state to it via the placeholder argument `&1`.