Small phrasing change for Heredocs
Issue
=====

When skimming the original phrasing of this sentence, I failed to
understand the sentence properly. I read the sentence

> Sigils also support heredocs, that is, triple double- or
single-quotes as separators

as

> Sigils also support heredocs, that is triple, double, or single
quotes as separators

The difference between these two sentences is only punctuation. The
good people of the Elixir Slack helped me to understand the true
intent of the sentence.

Proposal
========

I believe this phrasing is somewhat clearer. While I am a native
English speaker, I hope this rephrase is clearer to people who are not
native English speakers. I'd be interested to get feedback from the
community about this.