Fix version check command (#1673)
The current command proposed for checking the Erlang version is `erl -s halt`, but that makes `init` call the module `halt`, which does not exist, and it results in a crash. This may mislead users into thinking their Erlang version is not working properly.

In this change the module name is added so that it calls `erlang:halt` correctly. This results in no crash.