Remove links to Slack
It's not possible to join the Slack without being invited. Getting an
invite currently means that you actually have to go to one of the other
communities and then ask for an invite.

Examples of this:
- https://www.reddit.com/r/elixir/comments/129w7sf/how_to_join_elixir_slack/
- https://elixirforum.com/t/unable-to-join-the-slack-channel/56331
- https://www.reddit.com/r/elixir/comments/eawv49/cant_get_an_invitation_for_the_elixir_slack/

Pointing newcomers to ask questions in Slack is not great, since it
sends them down a rabbit hole of trying to find out how to even get
access to it.