Clarify Any Protocol implementation derivation
A small edit on the protocol derivation.
Changing "we should be fine with the implementation of Any" to
"should we be fine with the implementation of any" hopefully
better communicates that we can use the Any implementation of
the protocol if we wish to but not that we have to.